refactor(client): rename folders and package name from nft-client to client
