+++ packages/client/README.md
@@ -1,9 +1,9 @@
 ![Img](nft-client.png)
 [![License: CC BY-NC-SA 4.0](https://img.shields.io/badge/License-CC%20BY--NC--SA%204.0-lightgrey.svg)](https://creativecommons.org/licenses/by-nc-sa/4.0/)
 
-# NFT-Client
+# Client
 
-A Protokol module providing NFT API support for the ARK Core Blockchain Framework.
+A Protokol module providing protokol API support for the ARK Core Blockchain Framework.
 
 # License
