+++ packages/client/__tests__/mocks/base/assets.ts
