+++ packages/client/__tests__/mocks/exchange/configurations.ts
