+++ packages/client/__tests__/resources/base/assets.test.ts
