+++ packages/client/__tests__/resources/base/burns.test.ts
