+++ packages/client/__tests__/resources/base/configurations.test.ts
