+++ packages/client/__tests__/resources/exchange/configurations.test.ts
