+++ packages/client/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./nft-connection";
 export * as BaseResources from "./resources/base";
 export * as ExchangeResources from "./resources/exchange";
-export * as BaseResourcesTypes from "./resourcesTypes/base";
-export * as ExchangeResourcesTypes from "./resourcesTypes/exchange";
-export { PeerDiscovery } from "./peerDiscovery";
+export * as BaseResourcesTypes from "./resources-types/base";
+export * as ExchangeResourcesTypes from "./resources-types/exchange";
+export { PeerDiscovery } from "./peer-discovery";
 export { NFTConnectionManager } from "./nft-connection-manager";
