+++ packages/client/src/resources-types/base/assets.ts
