+++ packages/client/src/resources-types/base/burns.ts
