+++ packages/client/src/resources-types/base/configurations.ts
