+++ packages/client/src/resources-types/exchange/auctions.ts
