+++ packages/client/src/resources-types/exchange/trades.ts
