+++ packages/client/src/resources/base/burns.ts
@@ -1,6 +1,6 @@
 import { ApiResponse, ApiResponseWithPagination, Resource } from "@arkecosystem/client";
 
-import { AllBurnsQuery, Burns as BurnsResource } from "../../resourcesTypes/base";
+import { AllBurnsQuery, Burns as BurnsResource } from "../../resources-types/base";
 
 export class Burns extends Resource {
     public async all(query?: AllBurnsQuery): Promise<ApiResponseWithPagination<BurnsResource[]>> {
