+++ packages/client/src/resources/base/transfers.ts
@@ -1,7 +1,7 @@
 import { ApiResponse, ApiResponseWithPagination, Resource } from "@arkecosystem/client";
 
-import { Transfers as TransfersResource } from "../../resourcesTypes/base";
-import { AllTransfersQuery } from "../../resourcesTypes/base";
+import { Transfers as TransfersResource } from "../../resources-types/base";
+import { AllTransfersQuery } from "../../resources-types/base";
 
 export class Transfers extends Resource {
     public async all(query?: AllTransfersQuery): Promise<ApiResponseWithPagination<TransfersResource[]>> {
